requirements.txt
elevenlabs==2.8.1
kokoro==0.9.4
matplotlib==3.10.3
nemo_toolkit==2.3.1
numpy==2.3.2
openai==1.98.0
openwakeword==0.6.0
peft==0.15.2
PyAudio==0.2.14
pydub==0.25.1
pywebview==5.3.2
scipy==1.16.1
torch==2.7.0
transformers==4.46.3
websockets==13.1
